+++ kitty/window.py
@@ -436,7 +436,7 @@ class Window:
             self.screen.reset_callbacks()
         self.screen = None
 
-    def as_text(self, as_ansi=False, add_history=False, add_wrap_markers=False, alternate_screen=False):
+    def as_text(self, as_ansi=False, add_history=False, add_pager_history=False, add_wrap_markers=False, alternate_screen=False):
         lines = []
         add_history = add_history and not self.screen.is_using_alternate_linebuf() and not alternate_screen
         if alternate_screen:
@@ -446,6 +446,9 @@ class Window:
         f(lines.append, as_ansi, add_wrap_markers)
         if add_history:
             h = []
+            if add_pager_history:
+                # assert as_ansi and add_wrap_markers?
+                self.screen.historybuf.pagerhist_as_text(h.append)
             self.screen.historybuf.as_text(h.append, as_ansi, add_wrap_markers)
             lines = h + lines
         return ''.join(lines)
@@ -461,7 +464,7 @@ class Window:
     # actions {{{
 
     def show_scrollback(self):
-        data = self.as_text(as_ansi=True, add_history=True, add_wrap_markers=True)
+        data = self.as_text(as_ansi=True, add_history=True, add_pager_history=True, add_wrap_markers=True)
         data = data.replace('\r\n', '\n').replace('\r', '\n')
         lines = data.count('\n')
         input_line_number = (lines - (self.screen.lines - 1) - self.screen.scrolled_by)